uuid moved in string folder. uuid added to README.
