+++ src/string/uuid.ts
